RavenDB-24640 - add the BlittableJsonConverterWithMissingProperties
