+++ src/Raven.Client/Json/Serialization/NewtonsoftJson/Internal/BlittableJsonConverter.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using Newtonsoft.Json.Serialization;
 using Raven.Client.Documents.Session;
 using Sparrow.Json;
