+++ src/Raven.Client/Json/Serialization/NewtonsoftJson/Internal/BlittableJsonConverterBase.cs
@@ -17,8 +17,6 @@ namespace Raven.Client.Json.Serialization.NewtonsoftJson.Internal
     {
         protected readonly ISerializationConventions Conventions;
 
-        private Dictionary<object, Dictionary<object, object>>? _missingDictionary;
-
         protected BlittableJsonConverterBase(ISerializationConventions conventions)
         {
             Conventions = conventions ?? throw new ArgumentNullException(nameof(conventions));
@@ -280,45 +278,6 @@ namespace Raven.Client.Json.Serialization.NewtonsoftJson.Internal
                 default:
                     return null;
             }
-        }
-
-        public Dictionary<object, Dictionary<object, object>>? MissingProperties
-        {
-            get => _missingDictionary;
-            set => _missingDictionary = value;
-        }
-
-        public void Clear()
-        {
-            _missingDictionary?.Clear();
-        }
-
-        protected IEnumerable<KeyValuePair<object, object>> FillMissingProperties(object o)
-        {
-            if (_missingDictionary != null && _missingDictionary.TryGetValue(o, out var props))
-                return props;
-            return Enumerable.Empty<KeyValuePair<object, object>>();
-        }
-
-        protected void RegisterMissingProperties(object o, string id, object value)
-        {
-            if (Conventions.Conventions.PreserveDocumentPropertiesNotFoundOnModel == false ||
-                id == Constants.Documents.Metadata.Key)
-                return;
-
-            _missingDictionary ??= new Dictionary<object, Dictionary<object, object>>(ObjectReferenceEqualityComparer<object>.Default);
-
-            if (_missingDictionary.TryGetValue(o, out var dictionary) == false)
-            {
-                _missingDictionary[o] = dictionary = new Dictionary<object, object>();
-            }
-
-            dictionary[id] = value;
-        }
-
-        public void RemoveFromMissing<T>(T entity)
-        {
-            _missingDictionary?.Remove(entity);
         }
     }
 }
